+++ src/main/java/mcp/mobius/waila/api/ui/IDisplayHelper.java
@@ -8,7 +8,7 @@ public interface IDisplayHelper {
 
     void drawItem(MatrixStack matrixStack, int x, int y, ItemStack stack, float scale);
 
-    void drawGradientRect(MatrixStack matrixStack, int left, int top, int right, int bottom, int startColor, int endColor);
+    void drawGradientRect(MatrixStack matrixStack, float left, float top, float right, float bottom, int startColor, int endColor);
 
     void drawBorder(MatrixStack matrixStack, int minX, int minY, int maxX, int maxY, IBorderStyle border);
